backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.6
asyncpg>=0.28.0
python-multipart>=0.0.6
email-validator>=2.0.0
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
google-auth>=2.23.0
pydantic-settings>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
aiosqlite>=0.19.0
